Pull out importer definitions to _importer_impl.py
