+++ python/google/protobuf/importer.py
@@ -45,7 +45,7 @@ def protos(*args, **kwargs):
       'The protos function is only available on a 3.X interpreter.')
 
 
-if sys.version_info[0] > 2 and _api_implementation.Type() == 'cpp':
+if sys.version_info[0] > 2:
   try:
     from google.protobuf.pyext import _message as _pyext_message  # pylint: disable=g-import-not-at-top
   except ImportError as e:
@@ -57,181 +57,4 @@ if sys.version_info[0] > 2 and _api_implementation.Type() == 'cpp':
           'The protos function is only available when using the cpp implementation'
       )
   else:
-    from google.protobuf import message as _message  # pylint: disable=g-import-not-at-top
-    import contextlib  # pylint: disable=g-import-not-at-top
-    import importlib  # pylint: disable=g-import-not-at-top
-    import importlib.machinery  # pylint: disable=g-import-not-at-top
-    import os  # pylint: disable=g-import-not-at-top
-
-    from google.protobuf.internal import enum_type_wrapper  # pylint: disable=g-import-not-at-top
-    from google.protobuf import reflection as _reflection  # pylint: disable=g-import-not-at-top
-    from google.protobuf import symbol_database as _symbol_database  # pylint: disable=g-import-not-at-top
-
-    _sym_db = _symbol_database.Default()
-
-    _PROTO_MODULE_SUFFIX = '_pb2'
-
-    def _module_name_to_proto_file(module_name):
-      components = module_name.split('.')
-      proto_name = components[-1][:-1 * len(_PROTO_MODULE_SUFFIX)]
-      return os.path.sep.join(components[:-1] + [proto_name + '.proto'])
-
-    def _proto_file_to_module_name(proto_file):
-      components = proto_file.split(os.path.sep)
-      proto_base_name = os.path.splitext(components[-1])[0]
-      return '.'.join(components[:-1] +
-                      [proto_base_name + _PROTO_MODULE_SUFFIX])
-
-    @contextlib.contextmanager
-    def _augmented_syspath(new_paths):
-      original_sys_path = sys.path
-      if new_paths is not None:
-        sys.path = sys.path + new_paths
-      try:
-        yield
-      finally:
-        sys.path = original_sys_path
-
-    class ProtoLoader(importlib.abc.Loader):
-      """Instantiates a module equivalent to a _pb2.py file from a .proto."""
-
-      def __init__(self, module_name, protobuf_path):
-        self._module_name = module_name
-        self._protobuf_path = protobuf_path
-
-      def create_module(self, spec):
-        return None
-
-      def _generated_file_to_module_name(self, filepath):
-        components = filepath.split(os.path.sep)
-        return '.'.join(components[:-1] + [os.path.splitext(components[-1])[0]])
-
-      @staticmethod
-      def _register_message(sym_db, message_descriptor, module_name):
-        """Registers a message in the symbol database."""
-        nested_dict = {
-            '__module__': module_name,
-            'DESCRIPTOR': message_descriptor,
-        }
-        for nested_type_desc in message_descriptor.nested_types:
-          nested_type = ProtoLoader._register_message(sym_db, nested_type_desc,
-                                                      module_name)
-          nested_dict[nested_type_desc.name] = nested_type
-        for nested_enum_desc in message_descriptor.enum_types:
-          sym_db.RegisterEnumDescriptor(nested_enum_desc)
-        message_type = (
-            _reflection.GeneratedProtocolMessageType(message_descriptor.name,
-                                                     (_message.Message,),
-                                                     nested_dict))
-        sym_db.RegisterMessageDescriptor(message_descriptor)
-        sym_db.RegisterMessage(message_type)
-        return message_type
-
-      def exec_module(self, module):
-        """Instantiate a module identical to the generated version."""
-        file_descriptor = _pyext_message.FileDescriptor.FromFile(
-            self._protobuf_path.encode('ascii'),
-            [path.encode('ascii') for path in sys.path])
-        _sym_db.RegisterFileDescriptor(file_descriptor)
-        for dependency in file_descriptor.dependencies:
-          module_name = _proto_file_to_module_name(dependency.name)
-          if module_name not in sys.modules:
-            importlib.import_module(module_name)
-        setattr(module, 'DESCRIPTOR', file_descriptor)
-        for enum_name, enum_descriptor in file_descriptor.enum_types_by_name.iteritems(
-        ):
-          _sym_db.RegisterEnumDescriptor(enum_descriptor)
-          setattr(module, enum_name,
-                  enum_type_wrapper.EnumTypeWrapper(enum_descriptor))
-          for enum_value in enum_descriptor.values:
-            setattr(module, enum_value.name, enum_value.number)
-        for name, message_descriptor in file_descriptor.message_types_by_name.iteritems(
-        ):
-          message_type = ProtoLoader._register_message(_sym_db,
-                                                       message_descriptor,
-                                                       module.__name__)
-          _sym_db.RegisterMessageDescriptor(message_descriptor)
-          _sym_db.RegisterMessage(message_type)
-          setattr(module, name, message_type)
-
-    class ProtoFinder(importlib.abc.MetaPathFinder):
-      """MetaPathFinder that locates .proto files."""
-
-      def find_spec(self, fullname, path, target=None):
-        """Determines whether a .proto file can be loaded."""
-        del path
-        del target
-        filepath = _module_name_to_proto_file(fullname)
-        for search_path in sys.path:
-          try:
-            prospective_path = os.path.join(search_path, filepath)
-            os.stat(prospective_path)
-          except (FileNotFoundError, NotADirectoryError):
-            continue
-          else:
-            return importlib.machinery.ModuleSpec(
-                fullname, ProtoLoader(fullname, filepath))
-
-    def protos(protobuf_path, include_paths=None):  # pylint: disable=g-function-redefined, function-redefined
-      """Loads a module from a .proto file on disk.
-
-        This function is idempotent. Invoking it a second time will simply
-        return a module that has already been loaded. If the desired proto has
-        already been loaded from a "_pb2.py" file, this function will simply
-        return that module.
-
-        This function guarantees pointer equality for all types in the
-        dependency tree of an imported type. For example, suppose we have three
-        .proto files: A.proto, B.proto, and C.proto with the following contents:
-
-        A.proto:
-        ```
-        import "C.proto"
-
-        message AMessage {
-            CMessage c = 1;
-        }
-        ```
-
-        B.proto:
-        ```
-        import "C.proto"
-
-        message BMessage {
-            CMessage c = 1;
-        }
-        ```
-
-        C.proto:
-        ```
-        message CMessage {
-            int64 foo = 1;
-        }
-        ```
-
-        Then we guarantee that the following assertion will hold regardless of
-        the combination of generated code and runtime import used to load the
-        three modules:
-
-        ```
-        assert AMessage().c.__class__ is BMessage().c.__class__
-        ```
-
-      Args:
-        protobuf_path: A string representing the file path to the .proto file
-          on disk. (e.g. "google/protobuf/any.proto"). By default, all the
-          paths on sys.paths are searched for this file.
-        include_paths: An optional sequence of strings representing paths on
-          which to search for the .proto file to search in addition to the
-          entries in sys.paths.
-
-      Returns:
-        A module object identical to one that would be generated by protoc
-        and loaded with an `import foo_pb2` statement.
-      """
-      with _augmented_syspath(include_paths):
-        module_name = _proto_file_to_module_name(protobuf_path)
-        module = importlib.import_module(module_name)
-        return module
-
-    sys.meta_path.extend([ProtoFinder()])
+    from google.protobuf._importer_impl import protos
